Addressing Vercel error.
