+++ app/access/page.tsx
@@ -1,11 +1,25 @@
 
 import { prisma } from "@/lib/db";
-import { supabase } from "@/lib/supabase";
+import { createServerClient, type CookieOptions } from '@supabase/ssr'
+import { cookies } from 'next/headers'
+
 import NavBar from "@/components/NavBar";
 
 export default async function Dashboard() {
   
   // Fetch data
+  const cookieStore = cookies()
+  const supabase = createServerClient(
+    process.env.NEXT_PUBLIC_SUPABASE_URL!,
+    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
+    {
+      cookies: {
+        get(name: string) {
+          return cookieStore.get(name)?.value
+        },
+      },
+    }
+  )
   const { data } = await supabase.auth.getSession();
   const profile = await prisma.profile.findUnique({
     where: {id: data.session?.user.id},
